Start writing MOD files (0x10, 0x11, 0x12, 0x13)

+++ util/fstream_writer.hpp
@@ -25,7 +25,7 @@ public:
 
     inline void align(u32 amt)
     {
-        const u32 pos = tellp();
+        const u32 pos = static_cast<u32>(tellp());
         for (u32 i = 0; i < ((~(amt - 1) & (pos + amt - 1)) - pos); ++i) {
             writeU8(0);
         }
